+++ test/NumSharp.UnitTest/NpApiOverloads/NpApiOverloadTests_LogicManipulation.cs
@@ -543,7 +543,7 @@ public class NpApiOverloadTests_LogicManipulation
         var result = np.count_nonzero(a, axis: 0);
 
         await Assert.That(result).IsNotNull();
-        await Assert.That(result.shape).IsEquivalentTo(new int[] { 3 });
+        await Assert.That(result.shape).IsEquivalentTo(new long[] { 3 });
         await Assert.That(result.GetInt64(0)).IsEqualTo(1L);
         await Assert.That(result.GetInt64(1)).IsEqualTo(1L);
         await Assert.That(result.GetInt64(2)).IsEqualTo(2L);
@@ -557,7 +557,7 @@ public class NpApiOverloadTests_LogicManipulation
         var result = np.count_nonzero(a, axis: 1, keepdims: true);
 
         await Assert.That(result).IsNotNull();
-        await Assert.That(result.shape).IsEquivalentTo(new int[] { 2, 1 });
+        await Assert.That(result.shape).IsEquivalentTo(new long[] { 2, 1 });
     }
 
     #endregion
@@ -606,7 +606,7 @@ public class NpApiOverloadTests_LogicManipulation
         var result = np.dot(a, b);
 
         await Assert.That(result).IsNotNull();
-        await Assert.That(result.shape).IsEquivalentTo(new int[] { 2, 2 });
+        await Assert.That(result.shape).IsEquivalentTo(new long[] { 2, 2 });
         await Assert.That(result.GetInt32(0, 0)).IsEqualTo(19);
         await Assert.That(result.GetInt32(0, 1)).IsEqualTo(22);
         await Assert.That(result.GetInt32(1, 0)).IsEqualTo(43);
@@ -639,7 +639,7 @@ public class NpApiOverloadTests_LogicManipulation
         var result = np.matmul(a, b);
 
         await Assert.That(result).IsNotNull();
-        await Assert.That(result.shape).IsEquivalentTo(new int[] { 2, 2 });
+        await Assert.That(result.shape).IsEquivalentTo(new long[] { 2, 2 });
         await Assert.That(result.GetInt32(0, 0)).IsEqualTo(19);
         await Assert.That(result.GetInt32(0, 1)).IsEqualTo(22);
         await Assert.That(result.GetInt32(1, 0)).IsEqualTo(43);
@@ -662,7 +662,7 @@ public class NpApiOverloadTests_LogicManipulation
         var result = np.outer(a, b);
 
         await Assert.That(result).IsNotNull();
-        await Assert.That(result.shape).IsEquivalentTo(new int[] { 3, 3 });
+        await Assert.That(result.shape).IsEquivalentTo(new long[] { 3, 3 });
         await Assert.That(result.GetInt32(0, 0)).IsEqualTo(4);
         await Assert.That(result.GetInt32(0, 1)).IsEqualTo(5);
         await Assert.That(result.GetInt32(0, 2)).IsEqualTo(6);
@@ -706,7 +706,7 @@ public class NpApiOverloadTests_LogicManipulation
         var result = np.amin(a, axis: 0);
 
         await Assert.That(result).IsNotNull();
-        await Assert.That(result.shape).IsEquivalentTo(new int[] { 3 });
+        await Assert.That(result.shape).IsEquivalentTo(new long[] { 3 });
         await Assert.That(result.GetInt32(0)).IsEqualTo(0);
         await Assert.That(result.GetInt32(1)).IsEqualTo(0);
         await Assert.That(result.GetInt32(2)).IsEqualTo(2);
@@ -720,7 +720,7 @@ public class NpApiOverloadTests_LogicManipulation
         var result = np.amin(a, axis: 1, keepdims: true);
 
         await Assert.That(result).IsNotNull();
-        await Assert.That(result.shape).IsEquivalentTo(new int[] { 2, 1 });
+        await Assert.That(result.shape).IsEquivalentTo(new long[] { 2, 1 });
     }
 
     [Test]
@@ -745,7 +745,7 @@ public class NpApiOverloadTests_LogicManipulation
         var result = np.min(a, axis: 0);
 
         await Assert.That(result).IsNotNull();
-        await Assert.That(result.shape).IsEquivalentTo(new int[] { 3 });
+        await Assert.That(result.shape).IsEquivalentTo(new long[] { 3 });
         await Assert.That(result.GetInt32(0)).IsEqualTo(0);
         await Assert.That(result.GetInt32(1)).IsEqualTo(0);
         await Assert.That(result.GetInt32(2)).IsEqualTo(2);
@@ -769,7 +769,7 @@ public class NpApiOverloadTests_LogicManipulation
         var result = np.min(a, axis: 1, keepdims: true, dtype: null);
 
         await Assert.That(result).IsNotNull();
-        await Assert.That(result.shape).IsEquivalentTo(new int[] { 2, 1 });
+        await Assert.That(result.shape).IsEquivalentTo(new long[] { 2, 1 });
     }
 
     #endregion
@@ -813,7 +813,7 @@ public class NpApiOverloadTests_LogicManipulation
         var result = np.transpose(a);
 
         await Assert.That(result).IsNotNull();
-        await Assert.That(result.shape).IsEquivalentTo(new int[] { 3 });
+        await Assert.That(result.shape).IsEquivalentTo(new long[] { 3 });
     }
 
     [Test]
@@ -839,7 +839,7 @@ public class NpApiOverloadTests_LogicManipulation
         var result = np.dot(a, b);
 
         await Assert.That(result).IsNotNull();
-        await Assert.That(result.shape).IsEquivalentTo(new int[] { 2 });
+        await Assert.That(result.shape).IsEquivalentTo(new long[] { 2 });
         await Assert.That(result.GetInt32(0)).IsEqualTo(3);
         await Assert.That(result.GetInt32(1)).IsEqualTo(7);
     }
